+++ webpack/prod.config.js
@@ -1,6 +1,6 @@
 const common = require('./base.config.js')
 const merge = require('webpack-merge')
-const UglifyJSPlugin = require('uglifyjs-webpack-plugin')
+const UglifyJsPlugin = require('uglifyjs-webpack-plugin')
 
 module.exports = merge(common, {
   mode: 'production',
@@ -12,13 +12,29 @@ module.exports = merge(common, {
     // https://webpack.js.org/plugins/split-chunks-plugin/
     splitChunks: {
       chunks: 'initial' // default splitting, or 'all'
-    }
-  },
+    },
+    // inspired by and some issues to read more about treeshaking in webpack
+    // - https://github.com/gpspake/es6-todomvc/blob/c6e30fc6676ec362f3f4be8ec41df27c12007b8b/webpack.config.babel.js
+    // - https://stackoverflow.com/a/47675519/1238150 (treeshaking w/ change in .babelrc)
+    //   - see also: https://github.com/webpack-contrib/uglifyjs-webpack-plugin/issues/247#issuecomment-370911955
+    // - https://github.com/webpack-contrib/uglifyjs-webpack-plugin/issues/267 (Webpack 4 Issue)
+    // - https://github.com/webpack/webpack/issues/6992 (Webpack 4 Issue)
+    minimizer: [
+      new UglifyJsPlugin({
+        parallel: true,
+        sourceMap: false,
+        uglifyOptions: {
+          compress: {
+            drop_console: true,
+            dead_code: true,
+          },
+          output: {
+            beautify: false,
+            comments: false,
+          }
+        }
+      })
+    ]
+  }
 
-  plugins: [
-    new UglifyJSPlugin({
-      parallel: true,
-      sourceMap: false
-    })
-  ]
 })